Fix and document Grove 4-Digit Display and RGB Stick examples
- Migrated TM1637 and WS2813 examples to esp-hal 1.0.0
- Added tm1637-embedded-hal and esp-hal-smartled dependencies
- Added 5V wiring diagrams and warnings for 3.3V power issues
- Resolved clippy warnings across all examples

+++ examples/grove_oled.rs
@@ -210,7 +210,7 @@ fn main() -> ! {
                     .draw(&mut display)
                     .unwrap();
 
-                let progress_width = ((frame % 100) as u32 * 104) / 100;
+                let progress_width = ((frame % 100) * 104) / 100;
                 Rectangle::new(Point::new(12, 47), Size::new(progress_width, 8))
                     .into_styled(PrimitiveStyle::with_fill(BinaryColor::On))
                     .draw(&mut display)
